+++ packages/@sanity/form-builder/src/inputs/PortableText/Input.tsx
@@ -19,13 +19,12 @@ import {isEqual} from 'lodash'
 import {ChangeIndicatorWithProvidedFullPath} from '@sanity/base/components'
 import ActivateOnFocus from '../../components/ActivateOnFocus/ActivateOnFocus'
 import PatchEvent from '../../PatchEvent'
-import {BlockObject} from './Objects/BlockObject'
-import {InlineObject} from './Objects/InlineObject'
-import {EditObject} from './Objects/EditObject'
-import {Annotation} from './Text/Annotation'
+import {BlockObject} from './object/BlockObject'
+import {InlineObject} from './object/InlineObject'
+import {EditObject} from './object/EditObject'
+import {Annotation, TextBlock} from './text'
 import {RenderBlockActions, RenderCustomMarkers, ObjectEditData} from './types'
 import Editor from './Editor'
-import {TextBlock} from './Text/TextBlock'
 import {ExpandedLayer, Root} from './Input.styles'
 
 const ROOT_PATH = []
