+++ packages/@sanity/form-builder/src/inputs/PortableText/__workshop__/editObjects/EditObjectsStory.tsx
@@ -3,7 +3,7 @@ import {LayerProvider, PortalProvider, usePortal} from '@sanity/ui'
 import {useAction, useBoolean, useSelect} from '@sanity/ui-workshop'
 import React, {useMemo} from 'react'
 import FormBuilderContext from '../../../../FormBuilderContext'
-import {EditObject} from '../../Objects/EditObject'
+import {EditObject} from '../../object'
 import {ObjectEditData} from '../../types'
 import {resolveInputComponent, resolvePreviewComponent} from './formBuilder'
 import {schema, schemaType} from './schemaType'
