+++ packages/@sanity/form-builder/src/inputs/PortableText/object/index.ts
@@ -1,0 +1,1 @@
+export * from './EditObject'
