+++ packages/@sanity/form-builder/src/inputs/PortableText/text/index.ts
@@ -1,2 +1,4 @@
+export * from './Annotation'
+export * from './Decorator'
 export * from './TextBlock'
 export * from './textStyles'
