+++ packages/@sanity/form-builder/src/inputs/PortableText/toolbar/CustomIcon.tsx
