+++ packages/@sanity/form-builder/src/inputs/PortableText/toolbar/index.ts
@@ -1,0 +1,1 @@
+export * from './Toolbar'
